Essential, 5 Sequencing Computations, 5.1 Generics, 5.1.3.2 Working With Generic Types
