+++ src/main/scala/essentialScala/ex4/Ex4Traits.scala
@@ -1,4 +1,4 @@
-package essential_scala
+package essentialScala.ex4
 
 import java.util.Date
